pyproject.toml
[tool.poetry]
name = "eduplatform"
version = "0.1.0"
description = "EduPlatform - Ta'lim platformasi"
authors = ["Sizning Ismingiz <sizning@emailingiz.com>"]
readme = "README.md"
packages = [{include = "core"}, {include = "accounts"}, {include = "courses"}, {include = "payments"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.3"
drf-spectacular = "^0.27.1"
gunicorn = "^23.0.0"
whitenoise = "^6.7.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
Pillow = "^10.4.0"
celery = {extras = ["redis"], version = "^5.4.0"}
django-cors-headers = "^4.4.0"
dj-database-url = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
